// src/app/core/api/generated/schemas/returnReasonDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { ReturnReasonDtoDescription } from "./returnReasonDtoDescription";

export interface ReturnReasonDto {
  reason_id: number;
  reason_name: string;
  /** @nullable */
  description?: ReturnReasonDtoDescription;
  is_active: boolean;
  sort_order: number;
}